LPS-184817 Moves to util package and rename
