+++ modules/apps/layout/layout-api/bnd.bnd
@@ -11,7 +11,6 @@ Export-Package:\
 	com.liferay.layout.exception,\
 	com.liferay.layout.exporter,\
 	com.liferay.layout.friendly.url,\
-	com.liferay.layout.helper,\
 	com.liferay.layout.importer,\
 	com.liferay.layout.list.permission.provider,\
 	com.liferay.layout.list.retriever,\
